requirements.txt
bidict==0.22.0
cachelib==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Session==0.4.0
Flask-SocketIO==5.3.1
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.1
urllib3==1.26.12
websocket-client==1.4.2
Werkzeug==2.2.2
zope.event==4.5.0
zope.interface==5.5.1
Pillow==9.4.0
